Issue #5102: Move all parseXXX methods to the Parser class
