+++ src/test/java/com/puppycrawl/tools/checkstyle/AbstractTreeTestSupport.java
@@ -48,7 +48,7 @@ public abstract class AbstractTreeTestSupport extends AbstractPathTestSupport {
      * @throws Exception if exception occurs during verification.
      */
     protected static void verifyAst(String expectedTextPrintFileName, String actualJavaFileName,
-                                    AstTreeStringPrinter.PrintOptions withComments)
+                                    Parser.Options withComments)
             throws Exception {
         final String expectedContents = readFile(expectedTextPrintFileName);
 
@@ -62,7 +62,7 @@ public abstract class AbstractTreeTestSupport extends AbstractPathTestSupport {
     /**
      * Performs verification of the given text ast tree representation.
      * This implementation uses
-     * {@link AbstractTreeTestSupport#verifyAst(String, String, AstTreeStringPrinter.PrintOptions)}
+     * {@link AbstractTreeTestSupport#verifyAst(String, String, Parser.Options)}
      * method inside.
      * @param expectedTextPrintFileName expected text ast tree representation.
      * @param actualJavaFileName actual text ast tree representation.
@@ -71,7 +71,7 @@ public abstract class AbstractTreeTestSupport extends AbstractPathTestSupport {
     protected static void verifyAst(String expectedTextPrintFileName, String actualJavaFileName)
             throws Exception {
         verifyAst(expectedTextPrintFileName, actualJavaFileName,
-                AstTreeStringPrinter.PrintOptions.WITHOUT_COMMENTS);
+                Parser.Options.WITHOUT_COMMENTS);
     }
 
     /**
