+++ src/test/java/com/puppycrawl/tools/checkstyle/AstTreeStringPrinterTest.java
@@ -51,8 +51,7 @@ public class AstTreeStringPrinterTest extends AbstractTreeTestSupport {
     public void testParseFileThrowable() throws Exception {
         final File input = new File(getNonCompilablePath("InputAstTreeStringPrinter.java"));
         try {
-            AstTreeStringPrinter.printFileAst(input,
-                    AstTreeStringPrinter.PrintOptions.WITHOUT_COMMENTS);
+            AstTreeStringPrinter.printFileAst(input, Parser.Options.WITHOUT_COMMENTS);
             Assert.fail("exception expected");
         }
         catch (CheckstyleException ex) {
@@ -67,8 +66,7 @@ public class AstTreeStringPrinterTest extends AbstractTreeTestSupport {
     @Test
     public void testParseFile() throws Exception {
         verifyAst(getPath("ExpectedAstTreeStringPrinter.txt"),
-                getPath("InputAstTreeStringPrinterComments.java"),
-                AstTreeStringPrinter.PrintOptions.WITHOUT_COMMENTS);
+                getPath("InputAstTreeStringPrinterComments.java"), Parser.Options.WITHOUT_COMMENTS);
     }
 
     @Test
@@ -76,8 +74,7 @@ public class AstTreeStringPrinterTest extends AbstractTreeTestSupport {
         final FileText text = new FileText(
                 new File(getPath("InputAstTreeStringPrinterComments.java")).getAbsoluteFile(),
                 System.getProperty("file.encoding", StandardCharsets.UTF_8.name()));
-        final String actual = AstTreeStringPrinter.printAst(text,
-                AstTreeStringPrinter.PrintOptions.WITHOUT_COMMENTS);
+        final String actual = AstTreeStringPrinter.printAst(text, Parser.Options.WITHOUT_COMMENTS);
         final String expected = new String(Files.readAllBytes(Paths.get(
                 getPath("ExpectedAstTreeStringPrinter.txt"))), StandardCharsets.UTF_8);
 
@@ -88,7 +85,7 @@ public class AstTreeStringPrinterTest extends AbstractTreeTestSupport {
     public void testParseFileWithComments() throws Exception {
         verifyAst(getPath("ExpectedAstTreeStringPrinterComments.txt"),
                 getPath("InputAstTreeStringPrinterComments.java"),
-                AstTreeStringPrinter.PrintOptions.WITH_COMMENTS);
+                Parser.Options.WITH_COMMENTS);
     }
 
     @Test
@@ -121,21 +118,21 @@ public class AstTreeStringPrinterTest extends AbstractTreeTestSupport {
     public void testAstTreeBlockComments() throws Exception {
         verifyAst(getPath("ExpectedAstTreeStringPrinterFullOfBlockComments.txt"),
                 getPath("InputAstTreeStringPrinterFullOfBlockComments.java"),
-                AstTreeStringPrinter.PrintOptions.WITH_COMMENTS);
+                Parser.Options.WITH_COMMENTS);
     }
 
     @Test
     public void testAstTreeBlockCommentsCarriageReturn() throws Exception {
         verifyAst(getPath("ExpectedAstTreeStringPrinterFullOfBlockCommentsCR.txt"),
                 getPath("InputAstTreeStringPrinterFullOfBlockCommentsCR.java"),
-                AstTreeStringPrinter.PrintOptions.WITH_COMMENTS);
+                Parser.Options.WITH_COMMENTS);
     }
 
     @Test
     public void testAstTreeSingleLineComments() throws Exception {
         verifyAst(getPath("ExpectedAstTreeStringPrinterFullOfSinglelineComments.txt"),
                 getPath("InputAstTreeStringPrinterFullOfSinglelineComments.java"),
-                AstTreeStringPrinter.PrintOptions.WITH_COMMENTS);
+                Parser.Options.WITH_COMMENTS);
     }
 
 }
