+++ src/test/java/com/puppycrawl/tools/checkstyle/gui/MainFrameModelTest.java
@@ -197,7 +197,7 @@ public class MainFrameModelTest extends AbstractModuleTestSupport {
         }
         catch (CheckstyleException ex) {
             final String expectedMsg = String.format(Locale.ROOT,
-                    "NoViableAltException occurred while opening file %s.",
+                    "NoViableAltException occurred during the analysis of file %s.",
                     nonCompilableFile.getPath());
 
             assertEquals("Invalid exception message", expectedMsg, ex.getMessage());
