+++ src/test/java/com/puppycrawl/tools/checkstyle/internal/utils/TestUtil.java
@@ -19,28 +19,20 @@
 
 package com.puppycrawl.tools.checkstyle.internal.utils;
 
-import static com.puppycrawl.tools.checkstyle.TreeWalker.parseWithComments;
-
-import java.io.File;
-import java.io.IOException;
 import java.lang.reflect.Constructor;
 import java.lang.reflect.Field;
 import java.lang.reflect.Method;
 import java.lang.reflect.Modifier;
-import java.nio.charset.StandardCharsets;
 import java.util.Arrays;
 import java.util.Optional;
 import java.util.Set;
 import java.util.function.Predicate;
 
-import antlr.ANTLRException;
 import com.puppycrawl.tools.checkstyle.PackageNamesLoader;
 import com.puppycrawl.tools.checkstyle.PackageObjectFactory;
 import com.puppycrawl.tools.checkstyle.api.AbstractCheck;
 import com.puppycrawl.tools.checkstyle.api.CheckstyleException;
 import com.puppycrawl.tools.checkstyle.api.DetailAST;
-import com.puppycrawl.tools.checkstyle.api.FileContents;
-import com.puppycrawl.tools.checkstyle.api.FileText;
 
 public final class TestUtil {
 
@@ -179,20 +171,6 @@ public final class TestUtil {
             curNode = toVisit;
         }
         return Optional.ofNullable(curNode);
-    }
-
-    /**
-     * Parses Java source file. Results in AST which contains comment nodes.
-     * @param file file to parse
-     * @return DetailAST tree
-     * @throws NullPointerException if the text is null
-     * @throws IOException          if the file could not be read
-     * @throws ANTLRException       if parser or lexer failed
-     */
-    public static DetailAST parseFile(File file) throws IOException, ANTLRException {
-        final FileText text = new FileText(file.getAbsoluteFile(), StandardCharsets.UTF_8.name());
-        final FileContents contents = new FileContents(text);
-        return parseWithComments(contents);
     }
 
 }
