+++ src/test/java/com/puppycrawl/tools/checkstyle/xpath/XpathMapperTest.java
@@ -32,9 +32,9 @@ import java.util.List;
 import org.junit.Test;
 
 import com.puppycrawl.tools.checkstyle.AbstractPathTestSupport;
+import com.puppycrawl.tools.checkstyle.Parser;
 import com.puppycrawl.tools.checkstyle.api.DetailAST;
 import com.puppycrawl.tools.checkstyle.api.TokenTypes;
-import com.puppycrawl.tools.checkstyle.internal.utils.TestUtil;
 import net.sf.saxon.om.AxisInfo;
 import net.sf.saxon.om.Item;
 import net.sf.saxon.trans.XPathException;
@@ -527,7 +527,7 @@ public class XpathMapperTest extends AbstractPathTestSupport {
 
     private RootNode getRootNode(String fileName) throws Exception {
         final File file = new File(getPath(fileName));
-        final DetailAST rootAst = TestUtil.parseFile(file);
+        final DetailAST rootAst = Parser.parseFile(file);
         return new RootNode(rootAst);
     }
